fix(ReadsPipelineSpark): set numSlices to #files
This will generate one task for each input file split. With more than
one split in a task, the executor can run out of memory easily. Even
with only one split in each task, care must be taken so that each split
is small enough for in-memory processing because at least one task must
be completed before RDDs are created and spilled to the disk.